@@ -5,12 +5,13 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/futurehomeno/fimpgo"
-	"github.com/futurehomeno/fimpgo/utils"
-	log "github.com/sirupsen/logrus"
 	"io/ioutil"
 	"net/http"
 	"time"
+
+	"github.com/futurehomeno/fimpgo"
+	"github.com/futurehomeno/fimpgo/utils"
+	log "github.com/sirupsen/logrus"
 )
 
 type OAuth2TokenResponse struct {
@@ -73,7 +74,7 @@ func (oac *FhOAuth2Client) SetRefreshTokenApiUrl(refreshTokenApiUrl string) {
 	oac.refreshTokenApiUrl = refreshTokenApiUrl
 }
 
-//NewFhOAuth2Client implements OAuth client which communicates to 3rd party API over FH Auth proxy.
+// NewFhOAuth2Client implements OAuth client which communicates to 3rd party API over FH Auth proxy.
 func NewFhOAuth2Client(partnerName string, appName string, env string) *FhOAuth2Client {
 	client := &FhOAuth2Client{partnerName: partnerName, mqttServerURI: "tcp://localhost:1883", mqttClientID: "auth_client_" + appName}
 	if env == utils.EnvBeta {
@@ -124,7 +125,7 @@ func (oac *FhOAuth2Client) SetParameters(mqttServerUri, authCodeApiUrl, refreshT
 // ConfigureFimpSyncClient configures fimp sync client , which is used to obtain Hub token from cloud bridge.
 func (oac *FhOAuth2Client) ConfigureFimpSyncClient() error {
 	if oac.mqt == nil {
-		oac.mqt = fimpgo.NewMqttTransport(oac.mqttServerURI, oac.mqttClientID, "", "", true, 1, 1)
+		oac.mqt = fimpgo.NewMqttTransport(oac.mqttServerURI, oac.mqttClientID, "", "", true, 1, 1, nil)
 		err := oac.mqt.Start()
 		if err != nil {
 			log.Error("Error connecting to broker ", err)
@@ -146,17 +147,22 @@ func (oac *FhOAuth2Client) LoadHubTokenFromCB() error {
 		}
 	}
 	responseTopic := fmt.Sprintf("pt:j1/mt:rsp/rt:app/rn:%s/ad:1", oac.appName)
-	oac.syncClient.AddSubscription(responseTopic)
+	if err := oac.syncClient.AddSubscription(responseTopic); err != nil {
+		oac.syncClient.Stop()
+		oac.mqt.Stop()
+		return err
+	}
+
 	reqMsg := fimpgo.NewStringMessage("cmd.clbridge.get_auth_token", "clbridge", "", nil, nil, nil)
 	reqMsg.ResponseToTopic = responseTopic
 	var err error
 	var response *fimpgo.FimpMessage
-	for i := 0; i < oac.cbRetry; i++ {
+	for range oac.cbRetry {
 		response, err = oac.syncClient.SendFimp("pt:j1/mt:cmd/rt:app/rn:clbridge/ad:1", reqMsg, 5)
 		if err == nil {
 			break
 		}
-		log.Error("CB is not responding.Retrying")
+		log.Error("[edgeapp] CB is not responding.Retrying")
 		time.Sleep(time.Second * oac.cbRetryDelay)
 	}
 
