@@ -1,13 +1,14 @@
 package discovery
 
 import (
+	"testing"
+
 	"github.com/futurehomeno/fimpgo"
 	log "github.com/sirupsen/logrus"
-	"testing"
 )
 
 func SecondResponder() {
-	mqt := fimpgo.NewMqttTransport("tcp://localhost:1883", "fimpgotest-2", "", "", true, 1, 1)
+	mqt := fimpgo.NewMqttTransport("tcp://localhost:1883", "fimpgotest-2", "", "", true, 1, 1, nil)
 	err := mqt.Start()
 	if err != nil {
 		//t.Error("Error connecting to broker ",err)
@@ -32,7 +33,7 @@ func TestServiceDiscoveryResponder_Start(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
 	go SecondResponder()
 
-	mqt := fimpgo.NewMqttTransport("tcp://localhost:1883", "fimpgotest-1", "", "", true, 1, 1)
+	mqt := fimpgo.NewMqttTransport("tcp://localhost:1883", "fimpgotest-1", "", "", true, 1, 1, nil)
 	err := mqt.Start()
 	t.Log("Connected")
 	if err != nil {
@@ -54,7 +55,7 @@ func TestServiceDiscoveryResponder_Start(t *testing.T) {
 	responder.Start()
 
 	t.Log("Sending discovery request 1 ")
-	discoveredResource,_ := DiscoverResources(mqt, 2)
+	discoveredResource, _ := DiscoverResources(mqt, 2)
 	for _, r := range discoveredResource {
 		t.Log("Discovered resource = " + r.ResourceName)
 	}
@@ -63,7 +64,7 @@ func TestServiceDiscoveryResponder_Start(t *testing.T) {
 	}
 
 	t.Log("Sending discovery request 2 ")
-	discoveredResource,_ = DiscoverResources(mqt, 2)
+	discoveredResource, _ = DiscoverResources(mqt, 2)
 	for _, r := range discoveredResource {
 		t.Log("Discovered resource = " + r.ResourceName)
 	}
