@@ -24,7 +24,7 @@ func TestPrimeFimp_ClientApi_Update(t *testing.T) {
 	uuid := uuid.New().String()
 	validClientID := strings.ReplaceAll(uuid, "-", "")[0:22]
 
-	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1)
+	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1, nil)
 	err := mqtt.Start()
 	t.Log("Connected")
 	if err != nil {
@@ -76,7 +76,7 @@ func TestPrimeFimp_ClientApi_Notify(t *testing.T) {
 
 	validClientID := strings.ReplaceAll(uuid.New().String(), "-", "")[0:22]
 
-	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1)
+	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1, nil)
 	err := mqtt.Start()
 	t.Log("Connected")
 	if err != nil {
@@ -113,7 +113,7 @@ func TestPrimeFimp_SiteLazyLoading(t *testing.T) {
 
 	validClientID := strings.ReplaceAll(uuid.New().String(), "-", "")[0:22]
 
-	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1)
+	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1, nil)
 	err := mqtt.Start()
 	t.Log("Connected")
 	if err != nil {
@@ -137,8 +137,8 @@ func TestPrimeFimp_LoadStates(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
 
 	validClientID := strings.ReplaceAll(uuid.New().String(), "-", "")[0:22]
-	mqtt := fimpgo.NewMqttTransport(awsIotEndpoint, validClientID, brokerUser, brokerPass, true, 1, 1)
-	mqtt.ConfigureTls("awsiot.private.key","awsiot.crt","./datatools/certs",true)
+	mqtt := fimpgo.NewMqttTransport(awsIotEndpoint, validClientID, brokerUser, brokerPass, true, 1, 1, nil)
+	mqtt.ConfigureTls("awsiot.private.key", "awsiot.crt", "./datatools/certs", true)
 	mqtt.SetGlobalTopicPrefix(testSiteGuid)
 	err := mqtt.Start()
 	t.Log("Connected")
@@ -148,19 +148,19 @@ func TestPrimeFimp_LoadStates(t *testing.T) {
 
 	// Actual test
 	apiclientid := uuid.New().String()[0:12]
-	client := NewApiClient(apiclientid, mqtt, false,WithCloudService("test-proc-1"))
+	client := NewApiClient(apiclientid, mqtt, false, WithCloudService("test-proc-1"))
 	client.SetResponsePayloadType(fimpgo.CompressedJsonPayload)
 	state, err := client.GetState()
-	if err != nil || len(state.Devices)==0 {
+	if err != nil || len(state.Devices) == 0 {
 		t.Error("Cache is empty. Cache must contain data.")
-	}else {
-		t.Log("STATES - All Good .Number of states = ",len(state.Devices))
+	} else {
+		t.Log("STATES - All Good .Number of states = ", len(state.Devices))
 	}
 	shortcuts, err := client.GetShortcuts(false)
-	if err != nil || len(shortcuts)==0 {
+	if err != nil || len(shortcuts) == 0 {
 		t.Error("Cache is empty. Cache must contain data.")
-	}else {
-		t.Log("SHORTCUTS - All Good . Number of shortcuts = ",len(shortcuts))
+	} else {
+		t.Log("SHORTCUTS - All Good . Number of shortcuts = ", len(shortcuts))
 	}
 }
 
@@ -182,24 +182,24 @@ func TestPrimeFimp_LoadStatesWithConnPool(t *testing.T) {
 		StartFailRetryCount: 5,
 	}
 
-	connPool := fimpgo.NewMqttConnectionPool(3,5,20,time.Minute*3,transportConfigs,"lib_code_test_pool")
+	connPool := fimpgo.NewMqttConnectionPool(3, 5, 20, time.Minute*3, transportConfigs, "lib_code_test_pool")
 	connPool.Start()
 
 	var successCounter int
 
 	go func() {
 		for i := 0; i < 3; i++ {
-			connId, conn,err := connPool.BorrowConnection()
+			connId, conn, err := connPool.BorrowConnection()
 			if err != nil {
-				t.Fatal("Connection pool error , Err:",err.Error())
+				t.Fatal("Connection pool error , Err:", err.Error())
 			}
-			client := NewApiClient(validClientID, conn, false, WithCloudService("test-proc-1"),WithGlobalPrefix(testSiteGuid))
+			client := NewApiClient(validClientID, conn, false, WithCloudService("test-proc-1"), WithGlobalPrefix(testSiteGuid))
 			client.SetResponsePayloadType(fimpgo.CompressedJsonPayload)
 			state, err := client.GetState()
-			if err != nil || len(state.Devices)==0 {
+			if err != nil || len(state.Devices) == 0 {
 				t.Fatal("Cache is empty. Cache must contain data.")
-			}else {
-				t.Log("STATES - All Good .Number of states = ",len(state.Devices))
+			} else {
+				t.Log("STATES - All Good .Number of states = ", len(state.Devices))
 				successCounter++
 			}
 			connPool.ReturnConnection(connId)
@@ -225,11 +225,11 @@ func TestPrimeFimp_LoadStatesWithConnPool(t *testing.T) {
 		}
 	}()
 
-	time.Sleep(time.Second*10)
+	time.Sleep(time.Second * 10)
 
 	if successCounter != 6 {
 		t.Fatal("something went wrong")
-	}else {
+	} else {
 		t.Log("______ALL____GOOOD_______")
 	}
 
@@ -240,7 +240,7 @@ func TestPrimeFimp_ClientApi_Notify_With_Filter(t *testing.T) {
 
 	validClientID := strings.ReplaceAll(uuid.New().String(), "-", "")[0:22]
 
-	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1)
+	mqtt := fimpgo.NewMqttTransport(brokerUrl, validClientID, brokerUser, brokerPass, true, 1, 1, nil)
 	err := mqtt.Start()
 	t.Log("Connected")
 	if err != nil {
